+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v1.7.0
 
 + Multi-Currency support
++ CHANGE: A custom database name can now be set instead of being forced to use `totaleconomy` as the database
++ FIX: Account configuration file now properly saves when the server is stopped
 
 v1.6.0